+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## [0.2.0]
+
+- Null safety migration
+
+## [0.1.1-dev.2]
+
+- Dependencies updated
+
+## [0.1.1-dev.1]
+
+- Function `initRemoteConfForTest()` added
+
+## [0.1.0-dev.1]
+
+- Optional parameter `fetchTimeoutInSeconds` added in `init()`.
+- Dependencies updated.
+
 ## [0.0.1]
 
 - First release